requirements.txt
google-genai>=1.0.0
fpdf2>=2.7.9
Pillow>=10.0.0
requests>=2.31.0
httpx>=0.27.0
playwright>=1.44.0
pyyaml>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
